# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "High-performance RESTful backend API for the Timesheet Automation Portal, built with Go (Gin Engine) and PostgreSQL.\n- Automated Timesheet Generator: Multi-company Excel (.xlsx) and PDF generation (MII, SDD, Adidata, NTT) with day-fraction calculations, dynamic day trimming, formula preservation, and async SMTP delivery.\n- Dynamic Template Engine: Dedicated programmatic template builders alongside visual cell-mapping coordinate parsers integrated with Handsontable.\n- Dual Authentication \u0026 Security: JWT sessions with NIST SP 800-63B compliant Argon2id password hashing and WebAuthn FIDO2 Passkeys (biometric \u0026 hardware-bound).\n- Role-Based Access Control (RBAC): Admin and User role segregation, centralized account administration, and profile change approval workflows.\n- Web Push Notifications: Daily reminder scheduler pinned to 17:00 WIB (Asia/Jakarta) notifying active users via VAPID.\n- 3NF Relational Master Data: Normalized PostgreSQL database schema (Companies, Departments, Projects, Holidays, Activity Statuses) with foreign keys and automated SQL migrations.\n- API Versioning: Standardized RESTful endpoints routed under the `/api/v1/*` namespace.",
        "title": "Timesheet Automation Portal API",
        "termsOfService": "https://github.com/naufalzaid17/timesheet-generator",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "MIT"
...
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/push/subscribe": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registers or updates browser push notification subscription for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Push Notification"
                ],
                "summary": "Subscribe to Web Push notifications",
                "parameters": [
                    {
                        "description": "Push subscription payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SubscribeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid payload",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/push/test": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Dispatches an immediate test push notification to verify browser notification display.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Push Notification"
                ],
                "summary": "Send test push notification",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/push/unsubscribe": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes active push subscription for the browser, silencing timesheet reminder notifications.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Push Notification"
                ],
                "summary": "Unsubscribe from Web Push notifications",
                "parameters": [
                    {
                        "description": "Optional endpoint filter",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.UnsubscribeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid request or unauthorized"
                }
            }
        },
...
        "models.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "operation successful"
                }
            }
        },
...
        "models.PushKeyPayload": {
            "type": "object",
            "required": [
                "auth",
                "p256dh"
            ],
            "properties": {
                "auth": {
                    "type": "string"
                },
                "p256dh": {
                    "type": "string"
                }
            }
        },
...
        "models.SubscribeRequest": {
            "type": "object",
            "required": [
                "endpoint",
                "keys"
            ],
            "properties": {
                "endpoint": {
                    "type": "string",
                    "example": "https://fcm.googleapis.com/fcm/send/..."
                },
                "keys": {
                    "$ref": "#/definitions/models.PushKeyPayload"
                }
            }
        },
...
        "models.UnsubscribeRequest": {
            "type": "object",
            "properties": {
                "endpoint": {
                    "type": "string",
                    "example": "https://fcm.googleapis.com/fcm/send/..."
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter JWT token with format \"Bearer {token}\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}